pyproject.toml
[project]
name = "brainnet"
version = "0.1.0"
description = ""
authors = [
    {name = "Ea3124",email = "leesj6717@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "email-validator >=2.0.0,<3.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
ruff = "^0.11.10"
openai ="^1.79.0"
psycopg2-binary = "^2.9.9"
